requirements.txt
djangorestframework
markdown
coreapi
coreapi-cli
pytest
pytest-django
pytest-cov
WebTest
